move all implementation code to Message
